+++ apps/router-e2e/__e2e__/web-modal/app/nested/page.tsx
@@ -1,7 +1,5 @@
 import { Link } from 'expo-router';
-import { usePreventRemove } from 'expo-router/react-navigation';
-import { useState } from 'react';
-import { Alert, Pressable, ScrollView, StyleSheet, Text, View } from 'react-native';
+import { ScrollView, StyleSheet, Text, View } from 'react-native';
 
 const styles = StyleSheet.create({
   button: {
@@ -17,10 +15,6 @@ const styles = StyleSheet.create({
 });
 
 export default function Page() {
-  const [isPrevented, setIsPrevented] = useState(false);
-  usePreventRemove(isPrevented, () => {
-    Alert.alert('Prevented!', 'You cannot dismiss this modal right now.');
-  });
   return (
     <ScrollView style={{ padding: 16 }}>
       <Text
@@ -44,12 +38,6 @@ export default function Page() {
           Modal (transparent)
         </Link>
       </View>
-      <Pressable
-        onPress={() => {
-          setIsPrevented((p) => !p);
-        }}>
-        <Text style={styles.button}>{isPrevented ? 'Disable' : 'Enable'} preventRemove</Text>
-      </Pressable>
     </ScrollView>
   );
 }