+++ packages/expo-router/src/react-navigation/core/NavigationBuilderContext.tsx
@@ -9,17 +9,7 @@ export type ListenerMap = {
   focus: FocusedNavigationListener;
 };
 
-export type KeyedListenerMap = {
-  beforeRemove: ChildBeforeRemoveListener;
-};
-
 export type AddListener = <T extends keyof ListenerMap>(type: T, listener: ListenerMap[T]) => void;
-
-export type AddKeyedListener = <T extends keyof KeyedListenerMap>(
-  type: T,
-  key: string,
-  listener: KeyedListenerMap[T]
-) => void;
 
 export type ChildActionListener = (
   action: NavigationAction,
@@ -33,14 +23,11 @@ export type FocusedNavigationListener = <T>(callback: FocusedNavigationCallback<
   result: T;
 };
 
-export type ChildBeforeRemoveListener = (action: NavigationAction) => boolean;
-
 export type DispatchRoot = (
   action: NavigationAction,
   options?: {
     originKey?: string;
     suppressUnhandled?: boolean;
-    skipBeforeRemove?: boolean;
   }
 ) => boolean;
 
@@ -50,7 +37,6 @@ export type DispatchRoot = (
 export const NavigationBuilderContext = React.createContext<{
   onAction?: (action: NavigationAction, visitedNavigators?: Set<string>) => boolean;
   addListener?: AddListener;
-  addKeyedListener?: AddKeyedListener;
   dispatchRoot?: DispatchRoot;
   onDispatchAction: (action: NavigationAction, noop: boolean) => void;
   onOptionsChange: (options: object) => void;